:wrench:  Fixes #9 for TalkFutures
- Added dropdown select for Societies, ages, roles, and gender [previous commit] for TalkFutures deployment.
- Moved TalkFutures specific hard-coded data to Mixin for easy sharing.

This works, but there is a lot of duplicaed logic that could be abstracted.
